web/requirements.txt
aiofiles==0.3.2
aiokafka==0.4.0
async-timeout==3.0.0
blinker==1.4
click==6.7
gunicorn==19.9.0
h11==0.7.0
h2==3.0.1
hiredis==0.2.0
hpack==3.0.0
Hypercorn==0.1.0
hyperframe==5.1.0
itsdangerous==0.24
Jinja2==2.10
kafka-python==1.3.5
MarkupSafe==1.0
multidict==4.3.1
Quart==0.6.1
sortedcontainers==2.0.4
typing-extensions==3.6.5
uvloop==0.11.0
wsproto==0.11.0
